requirements.txt
Django==3.2.10
djangorestframework==3.12.4
pyTelegramBotAPI==4.1.1